requirements.txt
instagrapi
gspread
google-auth
google-generativeai
python-dotenv
pandas